backend/requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.4
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-vision==3.12.0
google-genai==1.61.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.7
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.8
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.8
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.2
numba==0.63.1
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.90
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.2.0
pikepdf==10.3.0
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.3.2
pypdf==6.6.2
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.31
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
pymongo==4.6.1
